+++ CHANGELOG.md
@@ -1,10 +1,15 @@
+### 2.1.0
+
+- Add Ancient Debris Ore
+- Fix French translation
+
 ### 2.0.0
 
 - Update to 1.21.8
 
 ### 1.1.0
 
-- Change smelting result to be feature equivalent to minecraft vanilla nether ore
+- Change smelting results to be feature equivalent to minecraft vanilla nether ore
 - Add missing block and item tags (`c:ore_rates/sparse` and `c:ores_in_ground/netherrack`)